Adaugare functionalitate folosire clapa MIDI in FL STUDIO (Native Instruments)
You can go to the Releases page, go to the latest release, download the "Native-Instruments-Komplete-Kontrol.zip" asset and extract the file on the FL Studio scripts folder, located on:

Windows: Documents\Image-Line\FL Studio\Settings\Hardware\
macOS: Documents/Image-Line/FL Studio/Settings/Hardware/
But if you want a more automated install and update process, you can use Git and a terminal:

Download and install Git

Open a terminal and run line by line:

Windows:

cd "%USERPROFILE%\Documents\Image-Line\FL Studio\Settings\Hardware"

git clone https://github.com/hobyst/flmidi-kompletekontrol.git --recurse-submodules "Native Instruments Komplete Kontrol"
macOS:

cd "~/Documents/Image-Line/FL Studio/Settings/Hardware"

git clone https://github.com/hobyst/flmidi-kompletekontrol.git --recurse-submodules "Native Instruments Komplete Kontrol"
Indifferently of whatever installation method you used, the folder structure inside the Hardware folder should look like this:

...

Native Instruments Komplete Kontrol/

.git/ (this folder will appear if you used the Git method)
lib/
_dummy_thread.py
nihia/
...
nihia.py
...
...
device_Komplete Kontrol.py
device_Komplete Kontrol (S-Series Custom MIDI Map).py
...
...

If inside the nihia folder there's no file called nihia.py, then you got the wrong asset file from the Releases page or you did git clone without the --recurse-submodules argument.

Assign the script to the device inside FL Studio
Connect your keyboard to your computer

Launch FL Studio

Go to Options > MIDI Settings. You'll find something like this:

There will be from 2 to 3 MIDI devices with the Komplete Kontrol name on it for each keyboard connected. The one we are interested on is the one with the "DAW" tag on it. This port is used by NI devices to send and recieve commands to and from the DAW software the user is using to enable the deep integration features.

Click on the DAW device entries on both Output and Input lists and assign them a port. It can be any port you want, but it has to be the same on both.

Click on the Enable button to tell FL Studio to recieve MIDI signals from the keyboard. Do it both on the MIDI and the DAW devices. On S-Series, you do it on both the DAW device and the one that is just named "Komplete Kontrol - X".

Finally, select the DAW device on both Input and Output lists, open the "Controller type" menu and select the "Native Instruments Komplete Kontrol (user)" script.

Now the device should have entered into deep integration mode, and the mixer should show up on the screen of the device.

Bonus for S-Series keyboards: There should be a second MIDI script on the list called "Native Instruments Komplete Kontrol (S-Series Custom MIDI Map) (user)". You can link this script to the MIDI port of your keyboard (named "Komplete Kontrol - X") to use the white buttons at the top of the screen when the keyboard is on MIDI mode to trigger different actions inside FL Studio (you need to be using the default MIDI template from the Komplete Kontrol software in order for this optional script to work):

Button	Action
1st	Creates a new pattern and asks for a name
2nd	Shows the playlist
3rd	Shows the channel rack
4th	Shows the piano roll
5th	Shows the full-screen plugin picker
6th	Shows the mixer
7th	Shows the MIDI settings
8th	Closes all windows
Troubleshooting
Powercycle your device: As simple or stupid it might appear to be, turning your device off and on might solve some problems in certain situations. Try it first before doing anything more if you have any problems:

S-Series: These ones have a dedicated power on/off button on the back. Use it to turn your device off and back on.

A-Series and M-Series: These power on and off only by the USB power. To turn off and on your device, just unplug it from your PC and then plug it back.

If this does nothing, try restarting your PC as well.

There's no DAW MIDI port (more typical on Windows, but might happen on macOS as well): The DAW port is a virtual MIDI port created by the drivers of the keyboard when they detect a Komplete Keyboard connected. If it doesn't appear then it might be due to a corrupt driver install or an error on the initialization of the driver. Try to reboot your PC. If that doesn't work, then:

Disconnect any NI device from your PC

Go to the Windows Settings (or the Control Panel) and uninstall all of the Komplete Kontrol named elements on the Apps and Features list:

Restart your computer

Run Native Access as an administrator and re-install Komplete Kontrol

Run Komplete Kontrol as an administrator to finalize the installation of the drivers

If you have an S-Series device, you might need to reinstall the standalone driver as well. You can download it from Native Instrument's drivers catalog

Restart your PC once again

If the problem persists, contact Native Instruments to get technical support

Change the device identifier inside the script:

Since the script is able to support multiple devices and some of them have differences in their behaviour, the script needs to know which kind of device you are using to make the controller behave properly. This is done automatically by the script by retrieving the name of the MIDI device is running on, so everything should work properly without the intervention of the user.

On the code, the script is set to interface with Komplete Kontrol S-Series MK2 devices, Maschine MK3 and Maschine Studio devices. If you are going to use one of devices, there shouldn't be any kind of problem. But if you are using a different device and you want to experiment with it or you are having problems, you can set it manually:

Go to the folder where the script is located

Open it using a text editor (like Notepad) or a code editor (like Visual Studio Code)

Scroll down a few lines and you will find a section defined by a stack of hashes # and titled "User-editable constants for script customization". There, you will find a few lines that look like this:

# Tells the script which kind of device are you using
# Possible values: "A_SERIES", "M_SERIES" or "S_SERIES"
DEVICE_SERIES = "S_SERIES"
The DEVICE_SERIES variable is the one that tells the script which device is interfacing. Change the "S_SERIES" text to be the identifier matching your device as described on the hashed text. If you are using a device that doesn't belong to the Komplete Kontrol family or you have doubts then this table might help you:

Device	DEVICE_SERIES ID
Komplete Kontrol S MK1 (not tested)	A_SERIES
Komplete Kontrol S MK2	S_SERIES
Komplete Kontrol M	A_SERIES or M_SERIES
Komplete Kontrol A	A_SERIES
Maschine MK3 (not tested)	S_SERIES
Maschine Studio (not tested)	S_SERIES
Other devices haven't been tested neither there's an actual idea on how they talk to the software, but trying each profile to see what works the best might be an option as well.

Once set to the right value, just save the file and FL Studio should catch up with the changes on the next launch.

If you are using more than one keyboard and they are from different series, on FL Studio you can go to View > Script output, click on the tab with the name of the device that is behaving badly (the main issue would be that the 4D Encoder doesn't act as expected) and modify the value of the DEVICE_SERIES variable at runtime by running this line on the interpreter:

DEVICE_SERIES = "<the identifier of your device>"
You'll have to do this every time you launch FL Studio and only for that device, but you can also duplicate the device_Komplete Kontrol.py file, open the copy, change the value of the variable and change the # name field on the first line of the script so it doesn't get confused with the other one. Then, link that second script to the device that wasn't behaving correctly and that's it.